...G...... [ZBXNEXT-9132] added version information to active check heartbeat request, use full version data in all agent2 requests/responses

+++ src/go/internal/agent/resultcache/diskcache.go
@@ -296,7 +296,7 @@ func (c *DiskCache) upload(u Uploader) (err error) {
 		Commands: cresults,
 		Session:  u.Session(),
 		Host:     u.Hostname(),
-		Version:  version.Short(),
+		Version:  version.Long(),
 		Variant:  agent.Variant,
 	}
