+++ src/go/internal/agent/resultcache/memorycache.go
@@ -58,7 +58,7 @@ func (c *MemoryCache) upload(u Uploader) (err error) {
 		Commands: c.cresults,
 		Session:  u.Session(),
 		Host:     u.Hostname(),
-		Version:  version.Short(),
+		Version:  version.Long(),
 		Variant:  agent.Variant,
 	}
